requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
mysqlclient==2.1.1
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
